requirements.txt
beeprint==2.4.7
certifi==2018.4.16
chardet==3.0.4
decorator==4.1.2
funcsigs==1.0.2
future==0.16.0
idna==2.7
jieba==0.39
mock==2.0.0
networkx==2.0
nltk==3.2.5
numpy==1.13.3
opencv-python==3.4.4.19
Pattern==2.6
pbr==3.1.1
Pillow==5.2.0
protobuf==3.5.1
requests==2.19.1
scipy==1.0.0
six==1.11.0
summa==0.0.7
tensorflow-gpu==1.0.1
tqdm==4.23.1
urllib3==1.23
urwid==2.0.1
